Revert "Separate project for timer class"
This reverts commit 65731d58174f7c32695908d98437be790d67ce5a.
