+++ .gitignore
@@ -461,7 +461,6 @@ _Pvt_Extensions
 
 # CLion
 .idea/
-**/cmake-build-debug/
 
 # VS Code
 .vscode/
