+++ dogm/demo/utils/include/timer.h
@@ -10,7 +10,7 @@
 #include <string>
 #include <vector>
 
-#include "IClock.h"
+#include "iclock.h"
 
 class Timer {
 public:
